changed data type to float64

+++ models/staging/raw/stg_raw__product.sql
@@ -10,7 +10,7 @@ renamed as (
 
     select
         products_id,
-        CAST(purchse_price as int64) AS purchase_price
+        CAST(purchse_price as float64) AS purchase_price
 
     from source
